Strip unit tests of module private functions
