+++ items-plan.md
@@ -4,6 +4,6 @@
 ## Open Questions
 
 * How do filters and so on map to shops? Do all items need to be defined in an item list's default? Could we just use a filter?
-
+* Do we actually need a collection config or is it a function parameter?
 
 ### Potential user journeys (might help)
