+++ src/core/loader/load-catalogue.js
